Refactor all parts of project
