+++ Backend/ETLLibrary/Interfaces/ISqlServerDatasetManager.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using ETLLibrary.Database;
+using ETLLibrary.Database.Models;
+using ETLLibrary.Database.Utils;
 
 namespace ETLLibrary.Interfaces
 {
