+++ Backend/ETLWebApp/Controllers/SqlServerDatasetController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using ETLLibrary.Authentication;
 using ETLLibrary.Database;
+using ETLLibrary.Database.Utils;
 using ETLLibrary.Interfaces;
 using Microsoft.AspNetCore.Mvc;
